// src/app/components/projects/project-list/productivity/prediction/model/act-enforcement-prediction.ts
export class PgActEnforcementPrediction {
  id: string;
  job_position: string;
  issuance_rate: number;
  issuance_unity: string;
  groupage: string;
  forecast_unity: string;
  forecast_deployed: number;
  project_id: number;
  project_name: string;
  forecast_per_unity: number;
  expected_unity: string;
  nbr_spaces_on_street_parking: number;
}